Added install prompt, fixed service worker & manifest
The WebApp can now be installed to desktop when the user clicks the install button on the index.html page (Firefox tested and working).

Due to PWA limitations, this can seemingly only be done once... ever. Users will need instructions on how to download and install the WebApp on a separate "help" page, if they miss the prompt.